Added some files that I have been working on
These are the files with all of the code ive been working on. The first one, SearchImageFromPoint.py takes a still image and, given a point, searches in an expanding loop for the character. The other Script uses that same process but does that every frame of a prerecorded video. Something that I tried with that file is not having it search the entire image. This is because processing the whole frame takes too long. These are the scripts that compleate the first two goals on the README file